rename #5: repository 폴더명 오타 수정
